requirements.txt
#========================= YOLOV8 ===============================
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0

#=========== Computer Vision/ Image Processing ==================
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

#=============== Data Processing & Visualization ================
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
azure-storage-blob>=12.19.0
python-dotenv>=1.0.0

#================ Notebook Utils ================================
ipykernel>=6.0.0
jupyter>=1.0.0
notebook>=7.0.0

#================ Utilities =====================================
tqdm>=4.65.0  # progress bar for processing videos
gdown>=5.2.0

#================ Testing ========================================
pytest>=7.4.0
pytest-cov>=4.1.0

#================ Code Quality ===================================
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

